Remove big integer (#166)
BigInt is only used in two places now and I think we can remove it:

### [VectorSessionToken comparison](https://github.com/Azure/azure-cosmos-js/blob/766720d8cd8726b9ac65eb6b0548e862f179f13f/src/session/VectorSessionToken.ts#L107) 

Replaced with a string comparison function. We shouldn't need to convert to BigInt.

### [Session Container](https://github.com/Azure/azure-cosmos-js/blob/master/src/session/sessionContainer.ts#L59) 

The `ownerID` string an 8 byte base64 string parsed by `ResourceId`. The first 4 bytes are the databse and the second 4 are the collection. I think ownerID is sufficient to use directly as it already uniquely identifies the database+collection. If it is not, we can use raw node Buffers to perform the logic [here](https://github.com/Azure/azure-cosmos-js/compare/master...remove-big-integer#diff-ac46183cc7f342585b6ce3920b0e414aL248). We'll also need a function for doing bitwise or on two buffers.