[Event Hubs] Clean up resources in finally block
We have negative tests that still hold onto resources after test
complete which causes Mocha to hang. This exposes an issue that
`HostContext` doesn't close client after an error is thrown.

This change ensures that in the error cases client is closed properly.
Mocha will also exit at the end so we can remove the hacky `--exit`
workaround that forces Mocha to exit after tests complete.