[JS] [Service Bus] package.json in the samples folder (#3118)
* [JS] [Service Bus] package.json in the samples folder

The samples in the `samples` folder fail to run because they require users to install `@azure/service-bus` while they are inside of the source code of `@azure/service-bus`. NPM does this distinction by finding the nearest `package.json`. A rather simple fix for this is to add a `package.json` inside of the `samples` folder. The package is named `@azure@/service-bus-samples` and is not published as part of this PR. The README was updated to reflect the changes.

Closes #3063