[Storage] Only iterate with an undefined marker or a non-empty marker
When using the paged iterator next() function and the continuation token, we get
an next marker of empty string. The service happily takes the empty string and
iterates from the beginning.

This change fixes it by only iterating when the marker is non-empty (more items
available) or undefined (starting from the beginning).

Samples are updated to check for `done`.