[Storage] config typedoc behavior
We still want to be able to jump to types that we own even they are from other
packages. For example, `PagedAsyncIterableIterator` is from `@core-paging`.
Setting `excludeExternals` to `false` and excluding files under `node_modules`
so that we could still have docs for types from other libraries in the repo.

Also exclude storage-specific files for generated operations.