Renaming keys => keyFilter, label => labelFilter
As part of a .net API review we discovered some inconsistent behavior
with how we handle filters, especially with regards to proper escaping.

This PR renames those fields and also makes them strings rather than
arrays while also documenting the format allowed for the underlying
string.

(also, fixing an odd issue where `nock` was failing. Adding the
dependency and removing a method that didn't need to be called)