[app-configuration] Use the x-ms-useragent header when in the browser (#6528)
Also, whitelisted the x-ms-useragent header for logging as well since 
it's the `User-Agent` equivalent header we use for browsers. (this 
isn't required to go out any time soon but can be useful for 
diagnostics).

This is the SDK-side of the change to fix #5810. The service side is tracked
in https://github.com/Azure/AppConfiguration/issues/217