[app-configuration] Allow the user to specify a user agent prefix (#6545)
This is a bit of a mucky PR that adds in support for the user to pass in their own user agent prefix

We're currently caught in between needing a custom throttling policy (to handle retries) which is different than the one in core-http. This "need" is making it difficult to use some of the standard conventions (like PipelineOptions) from core-http since they assume (probably rightfully) that you're buying into the entire set of policies wholesale.

So for now I'm just dodging that by making my existing AppConfigOptions for my constructor PipelineOptions compatible (but only for a small subset), in preparation for when core-http catches up and we can eliminate our custom throttling policy.