[KeyVault-Certificates] Poller public state implementation
This PR contains the implementation of a poller with a bag of public
properties. I believe the names I picked are decent, but I'll change
them if another name is suggested.

Please let me know if the current approach is not desireable too.

This also fixes the last remaining task from #6567 :
- getCertificateOperation should return a certificate with policy at
the end.

Therefore:

Fixes #6567