[Storage] fix nyc code coverage exclusion issue (#6609)
The code coverage report include data for dependencies (`dotenv` for
example). This is because we run the test on the one single bundled
file, while nyc exclusion works by not instrumenting imported or
required modules in the exclusion list.

This change run nyc on individual spec.js files instead. A dev
dependency `esm` is introduced so that nyc can load our test modules.

NodeJS part of #2820.