[eventhubs] Add a note about Roles for AAD Auth
Add a small note to the README.md which notes that when using AAD that
you need to assign a specific role to your principal in order to access
Event Hubs.

In addition, I changed some of the connection strings to be marked as
"code" so they render a little better by default.