[Storage][test] set larger timeout when uploading 257 MB
The duration of these tests normally is below 100 seconds but
occasionally goes over 120 seconds which is the timeout we specified
for mocha tests.  This change increases the timeout value to 180 which
hopefully should be large enough even for slow network, while still
small so not masking all the other potential test/product issues.