[event-hubs] Fixing bug in load balancing algorithm with even numbers of processors. (#6946)
Fixes #6439 - caused by an issue in our load balancing algorithm when dealing with
an event split of partitions to processors.

We were always assuming that we were allowed an "extra" partition, even
when the partitions would be evenly divisible by the number of processors.

NOTE: there is still a possibility of deadlock if a new processor comes online and
can't claim it's first partition. This is logged in issue #6945.