[storage][test] sharing the cached default HttpClient
Tests were using `newPipeline()` to create a pipeline then pass it to
the `BlobServiceClient` constructor.  `newPipeline()` was consider an
advanced usage (this can be debated) so we didn't try to offer a
HttpClient by default.

This change updates the blob tests to use the `BlobServiceClient`
directly so all clients can share the default HttpClient