[KeyVault] Tests to verify the clients' user agent (#7043)
Fixes #6840

What I'm trying to do is to have a test to check that the version in the package.json file is consistent with the version in the user agents. That's all the places where we change the versions prior to doing a new release.

To do this, I have to add the JSON plugin to rollup, to allow the test bundle to include the package.json file.

I also had to add the "resolveJsonModule": true flag to TypeScript's compiler.