Make ResourceResponse.resource optional
Currently the types indicate resource will never be undefined but that
is not the case when nothing is returned from a query in Cosmos. This
fix causes API to break in a sense that it may result in compile errors
however these compile errors may be masking bugs when the consumer is
using strict null checking.

Fixes #7819