[Storage] disable decompression of response when downloading (#7993)
* Add two failing tests due to decompression

* [Storage] disable decompression of response when downloading

The default behavior of node-fetch client is to decompress according
to `Accept-Encoding` when retrieving data. In blob storage it is
supported to set the content encoding of blobs so that browsers and
other clients can decompress accordingly. However we don't want this
behavior in our SDK libraries, because

- there's inconsistency between incoming data and outgoing data. Most
customers would not expect `client.download()` would give them
different data than what they upload in compressed format.

- Retry for partial response would not work on the decompressed response.

This change disables the decompression behavior for storage clients in
NodeJS.

Note that for browsers we don't have control and browser will always
honor the `Accept-Encoding` header and decompress accordingly.

Related to #6411.

* Update core-http dependency version to v1.1.0

* Document behavior differences between Node.js and browsers