[FormRecognizer] Use a static storage account for training tests
This change switches to use a persistent blob storage container
holding the training data needed by the live tests.

Urls to test documents also now point to blobs in a persistent blob
container in the same storage account.

Local copy of the training data are removed as they are no longer
needed.