Addressing the namespaces issue
After having a meeting with Jeff, we agreed that instead of disabling
these two ESLint rules, what made more sense was to chanthe namespaces
to become constant objects.

Even though this does change the API review file, the change will not
affect users, since a namespace is syntax sugar for an object.