[core-http] fix browser tests on Firefox
Issue 1) on Firefox user agent string contains more than two
fields:

  "core-http/1.1.3 OS/WindowsNT 10.0; Win64; x64"

while on Chrome it has just two:

  "core-http/1.1.3 OS/Win32"

Changed to verify it has at least two fields.

Issue 2) Most likely either Chrome has a bug, or behavior of dates
before January 1, 1970 is not well defined. For the date Tue, Jan 1
0001 00:00:00 GMT, Chrome sets the year to be 2001, while Firefox
keeps the year as 0001.

Changed to use another date in the test.

Issue 3) The error message when parsing xml with an input of undefined
or null is different between Chrome and Firefox.

Changed to pass on either one.