[Storage] Switch to userAgentPolicy() from core-http for telemetry
This change brings storage libraries closer to other libraries, in addition we
also get browser telemetry for free.

Note that this might break existing Storage customers in browsers due to the new
header we use to send user agent info, if the header is not allowed in their
CORS settings.