[core-http] set minRetryInterval to 50ms for exponential retry
In PR #9667 the initial retry delay was reduced, which result in the initial
delay being zero for exponential retry policy. While retrying immediately should
work, it caused some Key Vault tests to fail because they cancel the request in
1 millisecond then expect an `AbortError`. Immediate retry to a different error
than `AbortError`.

This change sets minRetryInterval to 50 so we don't retry immediately.