[Test Recorder] Fix an issue where not all recordings are saved
We use an async function to add a recorded request to an array then
save the array to files when recorder stops.  However we did not wait
for the async function to complete, which leads to recording missing
requests. This issue only happens in the browser case. In Node it is
done in synchronous way.

This change makes recorder to wait until all the recorded
requests are added before saving them.

All the tests have been updated to await for `recorder.stop()`.