[Test Recorder] Fix browser custom console logging when recording
In function `customConsoleLog()` we save `window.console.log` to a
local variable and use this variable to log to real console.  The
problem is that when `customConsoleLog()` is called multiple times,
`window.console.log` is no longer the original one, and we end up
nesting the modified `log` function again and again.

This PR saves the original console log function to a module level
variable instead so it's not wrapped multiple times.