[EventGrid] Use UTC Date in SAS Token test
The test code to generate a date was using `new Date()` which does a
local to UTC conversion. When run in CI we have a different local
timezone and so the expect data does not match the actual data.

Rewrite the test to use `Date.UTC` to construct a time that is in UTC.