StreamToString() does not corrupt images anymore.
The previous code corrupted files that aren't simple text files. The code corrupted all image files that I downloaded. The new code keeps the file in its original binary format.

A string does not have the same encoding as an image. To preserve the file in its proper binary code you have to keep it as a buffer. Call .toString() at the end of the download.