[EventGrid] Allow creating SAS tokens from browser
This change takes some of the utility code from the Cosmos SDK which
allows doing an SHA256 HMAC inside the browser (in addition to Node). In
order to support this, we need to make `generateSharedAccessSignature`
be async, since on the browser side some of the APIs we use are async
themselves.