[EventGrid] Move `generateSharedAccessSigniture`
Previously this was an instance method on `EventGridClient`. We've
decided that we would prefer it to be a top level function (where you
explicitly pass the endpoint and key) instead of an instance method that
re-uses these properties from the client (and fails when the client was
constructed with a shared access signiture itself, instead of a key).