[EventGrid] Tweak the consumer system type design
Instead of generating individual type guard methods for every system
event (which would lead to a bunch of exported functions which would
clutter up the documentation and like), we've adopted a model now where
we have a single `isSystemEvent` function.  This takes two arguments,
the first is a string which represents the type name of a System Event
supported by the SDK (and we have an union type with all these strings
in it) and the second is the CloudEvent or EventGridEvent object to
test.

In practice this means instead of:

```typescript
if (isContainerRegistryChartPushedEvent(event)) {
  // ...
} else if (isContainerRegistryChartDeletedEvent(event)) {
  // ...
}
```

You now write:

```typescript
if (isSystemEvent("Microsoft.ContainerRegistry.ChartPushed", event)) {
  // ...
} else if (isSystemEvent("Microsoft.ContainerRegistry.ChartDeleted", event)) {
  // ...
}
```

The `isSystemEvent` function is authored in such a way that auto
completion will show you the set of strings that could be passed into
isSystemEvent.