[service-bus] Receiver unification - BatchingReceiver (#10277)
As part of the ongoing enhancements to Track 2 we've been running into issues where code is getting duplicated between session and non-session (or worse, is drifting apart). 

This PR is first in a series to unify the implementations, starting with unifying BatchingReceiver and MessageSession to the same body of code.

This is tangentially related to #9829, which will get the same treatment but for StreamingReceiver and MessageSession.