[Key Vault Keys] testing RSA-HSM operations
Customer issue https://github.com/Azure/azure-sdk-for-js/issues/10352
highlighted a bug in our source code. The fact was that we weren't
testing RSA-HSM operations, and thus didn't catch the issue.

The bug consists of cryptography operations which should have been
technically allowed, but currently aren't due to outdated safety checks
and not because of an issue with the underlying technology.

This issue also highlighted the fact that our test resources for Key
Vault weren't considering having to run HSM operations, so the
test-resources.json file had to be updated to properly test this.

We still have to define how this is going to be released.

Fixes #10352