Fixing broken unit tests.
I had inadvertantly created an eternal call stack that would never end by causing negotiateClaim to cause continual calls to initLink and closeLink. This isn't even a recovery scenario as all the scenarios would eventually either fail by exhausting retries or because the link does eventually open.

In addition I added in some code and tests for a dead lock scenario that I observed. Basically we could end up with this call stack:

linkEntity.initLink()
  <enter lock>
  linkEntity.negotiateClaim()
    connectionContext.readyToOpenLink()
       refreshConnection() occurs
           calls linkEntity.closeLink()
              <attempts to enter lock and blocks>

The fix to this was two-fold:
1. Check the ready state of the connection before even entering into lock for initLink.
2. Within initLink, before we call negotiateClaim, do a status check to ensure that the connection is "ready". If it is not then go ahead and throw.