[service-bus, event-hubs] Allowing for a SAS key to be directly specified in a connection string. (#10951)
Allow for connection strings to be passed that contain a Shared Access Signature. 

This expands core-amqp so it can parse the connection string into a SharedKeyCredential. So connection strings of this form are valid:

`Endpoint=<endpoint>;SharedAccessSignature=SharedAccessSignature sr=<resource>...`

This results in a token that is not renewable and that returns the SAS verbatim.

Fixes #10832