[core-http] Treat '=' in query parameter values as valid
The '=' character is valid in query parameter values, it is also very
common in base64 encoding. Currently we treat it as invalid and any
parameter name/value pair with '=' in value gets ignore.

I have verified that both NodeJS and browsers' URL types allow '=' in
their `searchParameter` values.

This change remove the invalidation of this case in
`URLQuery.parse()` so that the parsing behavior is consistent with
NodeJS and browsers' URL.

Fixes #11014