[Event Grid] Improve typings
Split the input and output shapes for CloudEvent and EventGridEvent. On
input, we don't require either `id` or `eventTime` for EventGridEvent
or CloudEvent (since the client will use default values of a new UUID and the current
time), where as the outputs correctly reflect what properties will
always be set when consuming these events.