[event-hubs] Update docs to mention the checkpoint store updates made when the subscription is closed. (#11345)
As part of investigating #11316 we found it was non-obvious that we would be using the checkpoint store when closing. 

The reason we use it is when we close the subscription we also mark, in the checkpoint store, that the partitions are abandoned. This lets other consumers more quickly pick up abandoned partitions rather than forcing them to wait for an expiration interval and discover it.

Fixes #11316