[Service Bus] Fix test failures in the canary region - part 2 (#11597)
From https://github.com/Azure/azure-sdk-for-js/pull/11576
![image](https://user-images.githubusercontent.com/10452642/94872905-b18a2b00-0402-11eb-8339-8818586a594a.png)

### Background
We realized the max message size on the link has been increased with https://github.com/Azure/azure-sdk-for-js/pull/11588.

### Cause
The new test failures were caused since the messages that can be held/sent are have been increased through the new limit, more many messages being sent would mean that we'd receive all of them and verify per the test which would take more time and hence the timeouts.

### Fix in the test
Increase the message size and also create messages based on the maxMessageSize. This brings down the test time and would also be mindful of the potential size increase in the future.