Fixed the broken test. The original issue I had was that I wasn't re-adding the credits when recreating the link.
The second issue is that, in the browser, we actually do a context error (CloseEvent) reported, which causes our test (which expected no errors) to fail. From what I can tell this error seems reasonable in the browser and is directly related to websockets so for now I'm widening the test to accept it.