Fix the tests - the error reporting strategy inverted a bit.
Now we only report errors at the end of each cycle (ie, an entire retry<> call, not the individual operation).

Also, the operations now always throw a wrapper StreamingReceiverError that contains the actual error and is always set to retryable. This lets us preserve the proper error while still throwing an error that will cause retry<> to retry.