[core-http] de-serialize before checking internalError
Currently we check `parsedBody.error` or `parsedBody` for `code` and
`message` to use in the `RestError` to be thrown. However, we do this
on the raw object from parsing XML.

In the example added to the test, without this fix, the `parsedBody`
from xml parsing is
```js
{ Code: "ContainerAlreadyExists", Message: "The specified container already exists." }
```
but we are accessing the properties with lowercase `code` and `message`.

We should check on the object after de-serializing when we have error
object properly mapped already.