Some quality of life changes for those using MessagingError.
1. Make it so you can narrow the type ofr 'code' so you can get useful autocomplete information.

By default `isMessagingError` (which is new to this release) returns MessagingError's that are strictly typed (with an option to get the more loosely typed version that takes a 'string' for the .code field).

MessagingError itself (when no type parameter is specified) defaults to what it was previously - just a string.

So this change should be backwards compatible.

2. I've added in proper doc comments for the various error codes and alphabetized them. There's a _lot_ but at least this makes it easier to scan and find the one you want. Our samples in ServiceBus will clarify which of these codes are likely to be relevant for most users.