[Identity] VSCode credential default cloud and authority hosts (#11780)
If I understood this correctly, the solution to #11778 is to: default to AzureCloud on this credential.

This PR means that there will be no notion of "Azure" as a cloud name. However, if that credential exists (and only that exists) in the system, and no cloud name is officially set on the VSCode configuration files, it might still be taken, since we take whatever is available. That shouldn't be a problem on users that are authenticated through the VSCode extension, since they will have a concrete cloud name.

**IMPORTANT**:
After speaking with @g2vinay , he helped me realize that the VSCode credential still needs to map to the correct cloud based on the cloud name we get from the configurations! So I'll add that change to this PR.

Feedback appreciated.

Fixes #11778