Increment version for core-amqp v2 (#12267)
Next core-amqp release will be v2, updating files accordingly so that we can close #12266

Service Bus preview 8 is using core-amqp v2.0.0-beta.1 which has `MessageErrorCodes` and not `MessagingErrorCodes`. Updating SB to use `MessageErrorCodes` for now. Once SB preview 8 is released, we will update it to use v2.0.0 of core-amqp and revert this change