[event-hubs] fixes issue where processEvents ignores maxWaitTime after retryable disconnect event (#12280)
Fixes #12278 

## Description
This PR adds a check within the `PartitionPump`'s receive events loop to determine whether the receiver has been closed. If the receiver has been closed, it will create a new receiver using an event start position that matches the last event seen by the pump.

The receiver is explicitly closed when a disconnected event is received on the underlying AMQP connection, which causes calls to `receiveBatch` to immediately return any events it had collected up to this point. Note that once the receiver is closed, the receiver's onAmqpMessage handler is removed so it won't receive any additional events.

## Updates to testing
I manually tested the changes in this PR against the sample code in the linked issue.

I also updated the existing `disconnect` test to confirm that
- the `maxWaitTimeInSeconds` is honoured after a disconnected event is encountered.
- new events can be received on subsequent `processEvents` invocations.