[Service Bus] Remove Connection Context from ServiceBusMessage (#12398)
In #12216, we moved the code to settle a message and renew its lock from the ServiceBusMessageImpl to the ServiceBusReceiver. After this move, the only dependency left on the ConnectionContext in the ServiceBusMessageImpl is for the DataTransformer. This PR updates the ServiceBusMessage constructor to take the DataTransformer instead of the ConnectionContext and thus remove this dependency altogether. 

This closes #9944 as we no longer need to track the context or the receiver on the message.
This closes #10620 as well. Since the context is not being tracked on the message anymore, it does not matter that the data transformer is tracked on the context.