[Storage-Blob] Fix error listing blobs with "_" metadata key name
The cause of the issue is that xml parser in core-http was using
hard-coded "_" as the key to access parsed xml element content. Now
that core-http added support to customzie the XML_CHARKEY, fix on
Storage side simply pass "#" as the xml char key to use. Note that "#"
is not allowed in storage metadata keys so it's safe to use as the key
to access in xml parser.