[core-http] Support custom auth scopes (#12752)
Currently, there is no way to specify a custom scope to build the BearerTokenAuthenticationPolicy. So far `core-http` has been using the baseUri to build this scope. This generally works well for management plane libraries. However there are other scenarios where we may need to set a custom scope value, for example Synapse.

This PR introduces a new optional property to `ServiceClientOptions` called `authScope` which if present we'll use it as is for the scope, otherwise we fall back to the current behavior.