Adding in a tentative fix. This is theoretical at the moment but given what the symptoms in #12711 it seems to fit.
The basic problem would be - the user calls receiveMessages(5), for instance, and the timeout expires before they receive them all so we request a drain.

_Before_ the drain response comes we receive some of the other messages (after all, we requested 5). Those messages are then, via setTimeout, scheduled to be sent to our onMessage handler.

When the drain response arrives (prior to this fix) the drain would be processed immediately (unlike other messages the rhea-promise code does not call it with setTimeout), potentially bypassing any message events that are still waiting to be processed in the task queue.

By using setTimeout() here we use the same technique that the messages are scheduled by, ensuring that our resolve() is processed last.