Electron sample (#12960)
## What

Add an Electron sample that integrates with multiple Azure services.

## Why

As part of our investigation into how well our SDKs integrate into existing
frameworks we created a few sample apps that showcase how to use the
various SDKs. Adding a sample here would ensure that we have something
to reference and provide to folks who are interested in using our SDKs in Electron.