[Storage] retry on incomplete XML responses
When service times out (default max 30s) it terminates the connection
but current stable version of `node_fetch` doesn't report
error. Instead it returns the incomplete response which leads to XML
parse error.  It's unlikely that service would send back incomplete
response on purpose so it doesn't hurt to treat this error as a
`TIMEOUT` error and retry the request.

The deserialization policy factory needs to move below retry policy
factory so parse error from deserialization can be retried.